lmb-1460: move the date util methods to the date-manipulation util file and update all usage
