+++ app/components/date-input.js
@@ -7,8 +7,11 @@ import { restartableTask, timeout } from 'ember-concurrency';
 import moment from 'moment';
 
 import { INPUT_DEBOUNCE, NULL_DATE } from 'frontend-lmb/utils/constants';
-import { endOfDay } from 'frontend-lmb/utils/date-manipulation';
-import { isDateInRange, isValidDate } from 'frontend-lmb/utils/date';
+import {
+  endOfDay,
+  isDateInRange,
+  isValidDate,
+} from 'frontend-lmb/utils/date-manipulation';
 
 export default class DateInputComponent extends Component {
   elementId = `date-${guidFor(this)}`;
